Project Images
This is my implementation of the "BoomBox" which was pieced together using found parts and off the shelf hardware. 2x4 wood and PVC waste pipe make up the make components with a control box made from a discarded HP Plotter frame.  The glowing cylinder in the middle of image 432-434 is a standalone prop (A "Fuel Rod") - that can be sensed by the boombox and can sense the BoomBox and act out specific scenarios accordingly. It is a separate project I had before the BoomBox but made sense to integrate them, as it was fairly easy to accomplish.