Owner and Instruction Manual for BoomBox
Usage, maintenance, and construction guide including schematics and other relevant information. Note that this manual is tailored for the company I created the prop device for, D14 Airsoft. Do not re-use their name or content without their consent. (however the device itself I own all the IP for and have designated an MIT license for its use by the community.)